pyproject.toml
[project]
name = "checkout-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
